refactor(component): migrate Header to CSS Modules
