+++ src/components/Home/ToolFilters/ToolFilters.module.css
@@ -148,7 +148,3 @@ input[type="text"]::placeholder {
 	color: var(--text-muted);
 	margin-top: 0.25rem;
 }
-
-.white {
-	color: #fff;
-}
